# contracts/api/product/openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "Conexus Product API",
    "version": "4b-candidate",
    "description": "Canonical machine-readable HTTP wire for accepted Conexus Product operations. Product semantics remain owned by accepted 4A authority; this description may not invent operations, permissions, owners, or trust boundaries."
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Relative Product API root; deployment origin remains outside 4B."
    }
  ],
  "security": [
    {
      "ConexusSession": []
    }
  ],
  "paths": {
    "/api/control/access-context": {
      "get": {
        "summary": "GetControlPlaneAccessContext",
        "operationId": "GetControlPlaneAccessContext",
        "responses": {
          "200": {
            "description": "Current disclosable Control Plane access context resolved from server-owned authority.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "account",
                    "workspaces",
                    "projects"
                  ],
                  "properties": {
                    "account": {
                      "$ref": "#/components/schemas/AccountSummary"
                    },
                    "workspaces": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "workspaceId",
                          "name"
                        ],
                        "properties": {
                          "workspaceId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "pattern": ".*\\S.*",
                            "description": "Human-readable Workspace identity; presentation-only and never authorization."
                          }
                        }
                      }
                    },
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProjectSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          }
        }
      }
    },
...
    "/api/control/accounts": {
      "post": {
        "summary": "ProvisionAccount",
        "operationId": "ProvisionAccount",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapProvisionAccountRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conexus Account provisioned or the same admitted intake resolved to its already-created Account; Account presentation is returned without exposing provider authorization state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSummary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "409": {
            "$ref": "#/components/responses/Problem409"
          },
          "422": {
            "$ref": "#/components/responses/Problem422"
          }
        }
      }
    },
    "/api/control/workspaces/{workspaceId}/members": {
      "parameters": [
        {
          "$ref": "#/components/parameters/WorkspaceId"
        }
      ],
      "get": {
        "summary": "ListWorkspaceMembers",
        "operationId": "ListWorkspaceMembers",
        "responses": {
          "200": {
            "description": "Current Workspace roster. Members and invitations with their state are one server-resolved projection, and the caller's own role is stated so the browser can hide controls the server would refuse anyway.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceRoster"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          }
        }
      }
    },
    "/api/control/workspaces/{workspaceId}/members/{accountId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/WorkspaceId"
        },
        {
          "$ref": "#/components/parameters/AccountId"
        }
      ],
      "put": {
        "summary": "SetWorkspaceMemberRole",
        "operationId": "SetWorkspaceMemberRole",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "$ref": "#/components/schemas/WorkspaceRole"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The member now holds this role. Setting the same role again is the same answer."
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          },
          "409": {
            "$ref": "#/components/responses/Problem409"
          }
        }
      }
    },
    "/api/control/workspaces/{workspaceId}/invitations": {
      "parameters": [
        {
          "$ref": "#/components/parameters/WorkspaceId"
        }
      ],
      "post": {
        "summary": "InviteWorkspaceMember",
        "operationId": "InviteWorkspaceMember",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "email",
                  "role"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "minLength": 3,
                    "description": "Verified-email address the invited person must sign in with. It is the natural key of the invitation and never Account identity or authorization."
                  },
                  "role": {
                    "$ref": "#/components/schemas/WorkspaceRole"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current invitation for this exact Workspace and email. The pair is the natural key, so repeating the same request answers the same invitation rather than creating a second one.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInvitationEntry"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          },
          "422": {
            "$ref": "#/components/responses/Problem422"
          }
        }
      }
    },
    "/api/control/workspaces/{workspaceId}/roster/{entryKind}/{entryId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/WorkspaceId"
        },
        {
          "name": "entryKind",
          "in": "path",
          "required": true,
          "description": "Which kind of roster entry this request withdraws.",
          "schema": {
            "type": "string",
            "enum": [
              "member",
              "invitation"
            ]
          }
        },
        {
          "name": "entryId",
          "in": "path",
          "required": true,
          "description": "Untrusted reference to the exact roster entry; containment and authority are resolved server-side.",
          "schema": {
            "type": "string",
            "minLength": 1
          }
        }
      ],
      "delete": {
        "summary": "RemoveWorkspaceRosterEntry",
        "operationId": "RemoveWorkspaceRosterEntry",
        "responses": {
          "204": {
            "description": "The roster entry is gone. Removing a member withdraws every right that derived from that membership row."
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          },
          "409": {
            "$ref": "#/components/responses/Problem409"
          }
        }
      }
    },
...
    "/api/control/projects/{projectId}/application-access/{entryKind}/{entryId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        },
        {
          "name": "entryKind",
          "in": "path",
          "required": true,
          "description": "Which kind of application access entry this request withdraws.",
          "schema": {
            "type": "string",
            "enum": [
              "grant",
              "invitation"
            ]
          }
        },
        {
          "name": "entryId",
          "in": "path",
          "required": true,
          "description": "Untrusted reference to the exact access entry; containment and authority are resolved server-side.",
          "schema": {
            "type": "string",
            "minLength": 1
          }
        }
      ],
      "delete": {
        "summary": "RevokeApplicationAccessEntry",
        "operationId": "RevokeApplicationAccessEntry",
        "responses": {
          "204": {
            "description": "The access entry is gone. A revoked grant stops the person at their next request to the application."
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ConexusSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-conexus_session",
        "description": "Opaque server-owned Conexus iam.session carriage. Keycloak tokens, roles, groups, and organizations are not accepted as Product authorization. The server-issued cookie MUST be Secure, HttpOnly, SameSite=Lax, Path=/, and MUST NOT include a Domain attribute."
      }
    },
    "schemas": {
      "AccountSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accountId",
          "displayName"
        ],
        "properties": {
          "accountId": {
            "type": "string",
            "minLength": 1,
            "description": "Stable Conexus Account identity; presentation values never replace this identity."
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "pattern": ".*\\S.*",
            "description": "I&A-owned human presentation; never authentication or authorization."
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Optional presentation/contact data; never Account identity or authorization."
          }
        }
      },
      "ProjectSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "projectId",
          "workspaceId",
          "name",
          "archived"
        ],
        "properties": {
          "projectId": {
            "type": "string",
            "minLength": 1
          },
          "workspaceId": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "pattern": ".*\\S.*",
            "description": "Human-readable Project identity; not a stable identifier, uniqueness guarantee, or authorization source."
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
    },
...
    "responses": {
...
      "Problem401": {
        "description": "No valid current Conexus principal/session for the protected human surface.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem403": {
        "description": "Subject is legitimately disclosable to the current caller, but required action authority is denied.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    }
  }
}